Working on page count extraction
